# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cigar Helper API",
    "description": "API для управления хьюмидорами и сигарами",
    "contact": {
      "name": "Cigar Helper Team"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Profile/change-password": {
      "post": {
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Profile/avatar": {
      "post": {
        "tags": [
          "Profile"
        ],
        "summary": "Загрузка аватара (JPEG/PNG/WebP/AVIF/GIF), сохраняется как WebP до 384 px по длинной стороне.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyProfileDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profile"
        ],
        "summary": "Удалить загруженный аватар (внешний URL в поле не трогаем — только сброс ключа в хранилище).",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyProfileDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ReviewComments": {
      "get": {
        "tags": [
          "ReviewComments"
        ],
        "summary": "Список одобренных комментариев к обзору (только не удалённые обзоры).",
        "parameters": [
          {
            "name": "reviewId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewCommentDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ReviewComments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewCommentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewCommentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewCommentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/ReviewComments/{id}": {
      "delete": {
        "tags": [
          "ReviewComments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Reviews/{id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Получает детальную информацию об обзоре по его ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID обзора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Reviews"
        ],
        "summary": "Обновляет существующий обзор",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID обзора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Данные для обновления обзора",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "summary": "Мягкое удаление обзора автором (строка остаётся в БД, публично не отображается; восстановление — staff через `/api/admin/reviews/.../restore`).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID обзора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "required": [
          "confirmNewPassword",
          "currentPassword",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "currentPassword": {
            "minLength": 1,
            "type": "string"
          },
          "newPassword": {
            "maxLength": 100,
            "minLength": 6,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z0-9]).*$",
            "type": "string"
          },
          "confirmNewPassword": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ChangePasswordResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "retryAfterSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CigarCommentModerationStatus": {
        "enum": [
          "Pending",
          "Approved",
          "Rejected"
        ],
        "type": "string"
      },
...
      "CreateReviewCommentRequest": {
        "required": [
          "body",
          "reviewId"
        ],
        "type": "object",
        "properties": {
          "reviewId": {
            "type": "integer",
            "format": "int32"
          },
          "body": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CreateReviewImageRequest": {
        "required": [
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "imageUrl": {
            "minLength": 1,
            "type": "string"
          },
          "caption": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MyProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "isProfilePublic": {
            "type": "boolean"
          },
          "avatarUrl": {
            "type": "string",
            "description": "Внешний URL или относительный путь `/api/users/{id}/avatar` для загруженного файла.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ReviewCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "reviewId": {
            "type": "integer",
            "format": "int32"
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorUserId": {
            "type": "integer",
            "format": "int32"
          },
          "authorUsername": {
            "type": "string",
            "nullable": true
          },
          "isAuthorProfilePublic": {
            "type": "boolean"
          },
          "moderationStatus": {
            "$ref": "#/components/schemas/CigarCommentModerationStatus"
          }
        },
        "additionalProperties": false
      },
      "ReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "isAuthorProfilePublic": {
            "type": "boolean"
          },
          "userAvatarUrl": {
            "type": "string",
            "nullable": true
          },
          "cigarBaseId": {
            "type": "integer",
            "format": "int32"
          },
          "userCigarId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cigarName": {
            "type": "string",
            "nullable": true
          },
          "cigarBrand": {
            "type": "string",
            "nullable": true
          },
          "cigarCountry": {
            "type": "string",
            "nullable": true
          },
          "cigarLengthMm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cigarDiameter": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cigarWrapper": {
            "type": "string",
            "nullable": true
          },
          "cigarBinder": {
            "type": "string",
            "nullable": true
          },
          "cigarFiller": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewImageDto"
            },
            "nullable": true
          },
          "smokingExperience": {
            "type": "string",
            "nullable": true
          },
          "aroma": {
            "type": "string",
            "nullable": true
          },
          "taste": {
            "type": "string",
            "nullable": true
          },
          "construction": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "burnQuality": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "draw": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "nullable": true
          },
          "bodyStrengthScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "aromaScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pairingsScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "smokingDate": {
            "type": "string",
            "format": "date-time"
          },
          "smokingDurationMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "imageBytes": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "caption": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Role": {
        "enum": [
          "User",
          "Moderator",
          "Admin"
        ],
        "type": "string"
      },
...
      "UpdateProfileRequest": {
        "required": [
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "maxLength": 50,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "type": "string"
          },
          "email": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "isProfilePublic": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UpdateProfileResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "profile": {
            "$ref": "#/components/schemas/MyProfileDto"
          },
          "newToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateReviewRequest": {
        "required": [
          "content",
          "rating",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 3,
            "type": "string"
          },
          "content": {
            "minLength": 1,
            "type": "string"
          },
          "rating": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "smokingExperience": {
            "type": "string",
            "nullable": true
          },
          "aroma": {
            "type": "string",
            "nullable": true
          },
          "taste": {
            "type": "string",
            "nullable": true
          },
          "construction": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "burnQuality": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "draw": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "nullable": true
          },
          "smokingDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "smokingDurationMinutes": {
            "maximum": 720,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bodyStrengthScore": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "aromaScore": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pairingsScore": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "imagesToAdd": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateReviewImageRequest"
            },
            "nullable": true
          },
          "imageIdsToRemove": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}